Tests for empty array and object in section and each
One of these tests, 006-simple-section-011.json, does not pass with
the current lightncandy.  This is because {"foo": {}, "bar": []} has
the same output with json_decode($json, true) for both foo and bar.

To resolve the issue lightncandy would at some point in the future
need to handle popo's in the input data.